Added support for initializing pins at bootloader startup
Added support for initializing pins at bootloader startup,
this is different to "initial_pins" where it is only used
when the bootloader is running. Initializing the pins in
the bootloader is preferred because it is almost instant
rather than initializing it on the program where there
will be noticeable delay. Configuring is done on make menuconfi.